Refactor and take units into account in the contract
