+++ library/src/commonMain/kotlin/com/gu/recipe/template/TemplateParser.kt
@@ -1,0 +1,5 @@
+package com.gu.recipe.template
+
+data class ParsedTemplate(
+    val elements: List<TemplateElement>
+)
